Repair unit tests
