+++ tests/unit/testaddon.cpp
@@ -1147,11 +1147,11 @@ void TestAddon::telemetry_state_change() {
   // Event metrics can only record InvalidValue or InvalidOverflow errors
   // https://mozilla.github.io/glean/book/reference/metrics/event.html#recorded-errors
 
-  QCOMPARE(mozilla::glean::sample::addon_state_changed.testGetNumRecordedErrors(
-               VPNGlean::InvalidValue),
+  QCOMPARE(mozilla::glean::sample::addon_state_changed.numRecordedErrors(
+               EventMetric::InvalidValue),
            0);
-  QCOMPARE(mozilla::glean::sample::addon_state_changed.testGetNumRecordedErrors(
-               VPNGlean::InvalidOverflow),
+  QCOMPARE(mozilla::glean::sample::addon_state_changed.numRecordedErrors(
+               EventMetric::InvalidOverflow),
            0);
 
   // Upon creating the addon we expect two events to be recorded:
@@ -1175,11 +1175,11 @@ void TestAddon::telemetry_state_change() {
 
   // After disabling we expect a disabled state event to be recorded.
 
-  QCOMPARE(mozilla::glean::sample::addon_state_changed.testGetNumRecordedErrors(
-               VPNGlean::InvalidValue),
+  QCOMPARE(mozilla::glean::sample::addon_state_changed.numRecordedErrors(
+               EventMetric::InvalidValue),
            0);
-  QCOMPARE(mozilla::glean::sample::addon_state_changed.testGetNumRecordedErrors(
-               VPNGlean::InvalidOverflow),
+  QCOMPARE(mozilla::glean::sample::addon_state_changed.numRecordedErrors(
+               EventMetric::InvalidOverflow),
            0);
 
   auto postDisableValues =
