+++ vpnglean/src/event.cpp
@@ -11,9 +11,7 @@
 #include <QJsonObject>
 #include <QObject>
 #include <QPair>
-#if defined(UNIT_TEST)
-#  include <QJsonDocument>
-#endif
+#include <QJsonDocument>
 
 EventMetric::EventMetric(int id, EventMetricExtraParser* parser)
     : m_id(id), m_parser(parser) {}
@@ -60,9 +58,7 @@ void EventMetric::record(const EventMetricExtra& extras) {
   }
 }
 
-#if defined(UNIT_TEST)
-int32_t EventMetric::testGetNumRecordedErrors(
-    VPNGlean::ErrorType errorType) const {
+int32_t EventMetric::numRecordedErrors(ErrorType errorType) const {
 #  if not(defined(MZ_WASM) || defined(BUILD_QMAKE))
   return glean_event_test_get_num_recorded_errors(
       m_id, static_cast<int32_t>(errorType));
@@ -88,4 +84,3 @@ QList<QJsonObject> EventMetric::testGetValue(const QString& pingName) const {
   return QList();
 #  endif
 }
-#endif
